Samle kode i en fil
